Adding unique_filenames feature and misc cleanup
